+++ src/cli/printMaintainers.ts
@@ -1,0 +1,4 @@
+import {printMaintainers} from '../api/printMaintainers'
+import {runScript} from '../util/runScript'
+
+runScript(printMaintainers)